% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{as.static_path}
\alias{as.static_path}
\alias{as.static_path.static_path}
\alias{as.static_path.character}
\alias{as.static_path.default}
\title{convert an object to a static_path object}
\usage{
as.static_path(path)
\method{as.static_path}{static_path}(path)
\method{as.static_path}{character}(path)
\method{as.static_path}{default}(path)
}

\arguments{
\item{path}{the object to convert.}
}

\value{
a \code{static_path} object. character paths are converted to objects of class \code{static_path}; existing \code{static_path} objects are returned unchanged. objects of unsupported classes cause an error.
}

\description{
convert an object to a \code{static_path} object. This is primarily used internally, but can also be used by users to ensure that an object is a \code{static_path} object.
}

\keyword{internal}
