% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{as.staticPath}
\alias{as.staticPath}
\alias{as.staticPath.staticPath}
\alias{as.staticPath.character}
\alias{as.staticPath.default}
\title{Convert an object to a staticPath object}
\usage{
as.staticPath(path)
\method{as.staticPath}{staticPath}(path)
\method{as.staticPath}{character}(path)
\method{as.staticPath}{default}(path)
}

\arguments{
\item{path}{The object to convert.}
}

\value{
A \code{staticPath} object. Character paths are converted to objects of class \code{staticPath}; existing \code{staticPath} objects are returned unchanged. Objects of unsupported classes cause an error.
}

\description{
This function is used to convert an object to a \code{staticPath} object. It is primarily used internally, but can also be used by users to ensure that an object is a \code{staticPath} object.
}

\keyword{internal}
