% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{encode_uri}
\alias{encode_uri}
\alias{encode_uri_component}
\alias{decode_uri}
\alias{decode_uricomponent}
\title{URI encoding/decoding}
\usage{
encode_uri(value)
encode_uri_component(value)
decode_uri(value)
decode_uricomponent(value)
}

\arguments{
\item{value}{Character vector to be encoded or decoded.}
}

\value{
Encoded or decoded character vector of the same length as the input value. \\code\{decode_uri\} and \\code\{decode_uricomponent\} will return strings that are UTF-8 encoded.
}

\description{
Encodes/decodes strings using URI encoding/decoding in the same way that web browsers do. The precise behaviors of these functions can be found at developer.mozilla.org: \\href\{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encode_uri\}\{encode_uri\}, \\href\{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/encode_uri_component\}\{encode_uri_component\}, \\href\{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/decode_uri\}\{decode_uri\}, \\href\{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/decode_uricomponent\}\{decode_uricomponent\}

Intended as a faster replacement for \code{\link[utils:URLencode]{utils::URLencode()}} and \code{\link[utils:URLdecode]{utils::URLdecode()}}. encode_uri differs from encode_uri_component in that the former will not encode reserved characters: \\code\{;,/?:@@&=+$\}

decode_uri differs from decode_uricomponent in that it will refuse to decode encoded sequences that decode to a reserved character. (If in doubt, use decode_uricomponent.)

For \\code\{encode_uri\} and \\code\{encode_uri_component\}, input strings will be converted to UTF-8 before URL-encoding.
}

