% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in httpserver-package.R
\name{httpserver-package}
\alias{httpserver-package}
\title{HTTP and WebSocket server library}
\description{
Low-level socket and protocol support for handling HTTP and WebSocket requests. It is primarily intended as a building block for other R packages.
}

