% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in socket.R
\name{interrupt}
\alias{interrupt}
\title{interrupt httpuv runloop}
\usage{
interrupt()
}

\value{
no return value, called for its side effect of pausing the request processing loop.
}

\description{
interrupts the currently running httpuv runloop, meaning \code{\link[=run_server]{run_server()}} or \code{\link[=service]{service()}} will return control back to the caller and no further tasks will be processed until those methods are called again. note that this may cause in-process uploads or downloads to be interrupted in mid-request.
}

