% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4r.R
\name{ip_family}
\alias{ip_family}
\title{Check whether an address is IPv4 or IPv6}
\usage{
ip_family(ip)
}

\arguments{
\item{ip}{A single string representing an IP address.}
}

\value{
For IPv4 addresses, \\code\{4\}; for IPv6 addresses, \\code\{6\}. If the address is neither, \\code\{-1\}.
}

\description{
Given an IP address, this checks whether it is an IPv4 or IPv6 address.
}

\examples{
 ip_family("127.0.0.1")   # 4
 ip_family("500.0.0.500") # -1
 ip_family("500.0.0.500") # -1

 ip_family("::")          # 6
 ip_family("::1")         # 6
 ip_family("fe80::1ff:fe23:4567:890a") # 6
}

