% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{list_servers}
\alias{list_servers}
\title{list all running httpserver servers}
\usage{
list_servers()
}

\value{
a list of currently running server objects. the list is empty when no servers are running.
}

\description{
this returns a list of all running httpserver server applications.
}

