% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{log_level}
\alias{log_level}
\title{get and set logging level}
\usage{
log_level(level = NULL)
}

\arguments{
\item{level}{the logging level. must be one of \code{NULL}, \code{"off"}, \code{"error"}, \code{"warn"}, \code{"info"}, or \code{"debug"}. if \code{NULL} (the default), then this function simply returns the current logging level.}
}

\value{
if \code{level=NULL}, then this returns the current logging level. if \code{level} is any other value, then this returns the previous logging level, from before it is set to the new value.
}

\description{
the logging level for httpserver can be set to report differing levels of information. possible logging levels (from least to most information reported) are: \code{"off"}, \code{"error"}, \code{"warn"}, \code{"info"}, or \code{"debug"}. the default level is \code{error}.
}

\keyword{internal}
