% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{PipeServer}
\alias{PipeServer}
\title{PipeServer class}
\description{
This class represents a server running one application that listens on a
named pipe.
}

\seealso{
[Server()] and [WebServer()].
}

\keyword{internal}
