% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in socket.R
\name{raw_to_base64}
\alias{raw_to_base64}
\title{convert raw vector to base64-encoded string}
\usage{
raw_to_base64(x)
}

\arguments{
\item{x}{a raw vector.}
}

\value{
a single-element character vector containing the base64 encoding of \code{x}.
}

\description{
converts a raw vector to its base64 encoding as a single-element character vector.
}

\examples{
set.seed(100)
result <- raw_to_base64(as.raw(runif(19, min = 0, max = 256)))
stopifnot(identical(result, "TkGNDnd7z16LK5/hR2bDqzRbXA=="))
}

