% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in socket.R
\name{run_server}
\alias{run_server}
\title{run a server}
\usage{
run_server(host, port, app)
}

\arguments{
\item{host}{a string that is a valid ipv4 or ipv6 address that is owned by this server, which the application will listen on. \code{"0.0.0.0"} represents all ipv4 addresses and \code{"::/0"} represents all ipv6 addresses.}

\item{port}{a number or integer that indicates the server port that should be listened on. note that on most unix-like systems including linux and mac_os, port numbers smaller than 1024 require root privileges.}

\item{app}{a collection of functions that define your application. see \code{\link[=start_server]{start_server()}}.}
}

\value{
normally does not return; after interruption, returns the logical value \code{TRUE} from \code{\link[=service]{service()}}. the server is stopped as an exit side effect.
}

\description{
this is a convenience function that provides a simple way to call \code{\link[=start_server]{start_server()}}, \code{\link[=service]{service()}}, and \code{\link[=stop_server]{stop_server()}} in the correct sequence. it does not return unless interrupted or an error occurs.
}

\details{
if you have multiple hosts and/or ports to listen on, call the individual functions instead of \code{run_server}.
}

\seealso{
\code{\link[=start_server]{start_server()}}, \code{\link[=service]{service()}}, \code{\link[=stop_server]{stop_server()}}
}

\examples{
# a very basic application. run_server() blocks until interrupted, so
# schedule an interrupt() call to let this example return.
later2::later(interrupt, delay = 1)
run_server(
  "0.0.0.0", 5000,
  list(
    call = function(req) {
      list(
        status = 200L,
        headers = list(
          "content-type" = "text/html"
        ),
        body = "hello world!"
      )
    }
  )
)
}

