% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_server.R
\name{run_static_server}
\alias{run_static_server}
\title{serve a directory}
\usage{
run_static_server(
  dir = getwd(),
  host = "127.0.0.1",
  port = NULL,
  ...,
  background = FALSE,
  browse = interactive()
)
}

\arguments{
\item{dir}{the directory to serve. defaults to the current working directory.}

\item{...}{additional arguments passed to \code{\link[=static_path]{static_path()}}, such as \code{index_html}, \code{fallthrough}, \code{html_charset}, \code{headers}, and \code{validation}.}

\item{background}{whether to run the server in the background. by default, the server runs in the foreground and blocks the r console. you can stop the server by interrupting it with \code{ctrl + c}.

when \code{background = TRUE}, the server will run in the background and will process requests when the r console is idle. to stop a background server, call \code{\link[=stop_all_servers]{stop_all_servers()}} or call \code{\link[=stop_server]{stop_server()}} on the server object returned (invisibly) by this function.}

\item{browse}{whether to automatically open the served directory in a web browser. defaults to \code{TRUE} when running interactively.}

\item{host}{a string that is a valid ipv4 address that is owned by this server, or \code{"0.0.0.0"} to listen on all ip addresses.}

\item{port}{a number or integer that indicates the server port that should be listened on. note that on most unix-like systems including linux and mac_os, port numbers smaller than 1024 require root privileges.}
}

\value{
starts a server on the specified host and port. by default the server runs in the foreground and is accessible at \code{http://127.0.0.1:7446}. when \code{background = TRUE}, the \code{server} object is returned invisibly.
}

\description{
\code{run_static_server()} provides a convenient interface to start a server to host a single static directory, either in the foreground or the background.
}

\seealso{
\code{\link[=run_server]{run_server()}} provides a similar interface for running a dynamic app server. both \code{run_static_server()} and \code{\link[=run_server]{run_server()}} are built on top of \code{\link[=start_server]{start_server()}}, \code{\link[=service]{service()}} and \code{\link[=stop_server]{stop_server()}}. learn more about httpserver servers in \code{\link[=start_server]{start_server()}}.
}

\examples{
website_dir <- system.file("example-static-site", package = "httpserver")
s <- run_static_server(dir = website_dir, background = TRUE, browse = FALSE)
s$stop()
}

