% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{Server}
\alias{Server}
\title{Server class}
\description{
The `Server` class is the parent class for [WebServer()] and
[PipeServer()]. This class defines an interface and is not meant to
be instantiated.
}

\seealso{
[WebServer()] and [PipeServer()].
}

\keyword{internal}
