% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{static_path}
\alias{static_path}
\alias{exclude_static_path}
\alias{print.static_path}
\alias{format.static_path}
\title{create a static_path object}
\usage{
static_path(
  path,
  index_html = NULL,
  fallthrough = NULL,
  html_charset = NULL,
  headers = NULL,
  validation = NULL
)
exclude_static_path()
\method{print}{static_path}(x, ...)
\method{format}{static_path}(x, ...)
}

\arguments{
\item{path}{the local path.}

\item{x}{a \code{static_path} object.}

\item{...}{further arguments passed to or from other methods (currently unused).}

\item{index_html}{if an index.html file is present, should it be served up when the client requests the static path or any subdirectory?}

\item{fallthrough}{with the default value, \code{FALSE}, if a request is made for a file that doesn't exist, then httpserver will immediately send a 404 response from the background i/o thread, without needing to call back into the main r thread. this offers the best performance. if the value is \code{TRUE}, then instead of sending a 404 response, httpserver will call the application's \code{call} function, and allow it to handle the request.}

\item{html_charset}{when html files are served, the value that will be provided for \code{charset} in the content-type header. for example, with the default value, \code{"utf-8"}, the header is \code{content-type: text/html; charset=utf-8}. if \code{""} is used, then no \code{charset} will be added in the content-type header.}

\item{headers}{additional headers and values that will be included in the response.}

\item{validation}{an optional validation pattern. presently, the only type of validation supported is an exact string match of a header. for example, if \code{validation} is \code{'"abc" = "xyz"'}, then http requests must have a header named \code{abc} (case-insensitive) with the value \code{xyz} (case-sensitive). if a request does not have a matching header, than httpserver will give a 403 forbidden response. if the \code{character(0)} (the default), then no validation check will be performed.}
}

\value{
an object of class \code{static_path}, consisting of the normalized local path and a \code{static_path_options} object describing how it is served.
}

\description{
the \code{static_path} function creates a \code{static_path} object. note that if any of the arguments (other than \code{path}) are \code{NULL}, then that means that for this particular static path, it should inherit the behavior from the static_path_options set for the application as a whole.
}

\details{
the \code{exclude_static_path} function tells the application to ignore a particular path for static serving. this is useful when you want to include a path for static serving (like \code{"/"}) but then exclude a subdirectory of it (like \code{"/dynamic"}) so that the subdirectory will always be passed to the r code for handling requests. \code{exclude_static_path} can be used not only for directories; it can also exclude specific files.
}

\seealso{
\code{\link[=static_path_options]{static_path_options()}}.
}

