% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{static_path_options}
\alias{static_path_options}
\alias{print.static_path_options}
\alias{format.static_path_options}
\title{create options for static paths}
\usage{
static_path_options(
  index_html = TRUE,
  fallthrough = FALSE,
  html_charset = "utf-8",
  headers = list(),
  validation = character(0),
  exclude = FALSE
)
\method{print}{static_path_options}(x, ...)
\method{format}{static_path_options}(x, ...)
}

\arguments{
\item{index_html}{if an index.html file is present, should it be served up when the client requests the static path or any subdirectory?}

\item{fallthrough}{with the default value, \code{FALSE}, if a request is made for a file that doesn't exist, then httpserver will immediately send a 404 response from the background i/o thread, without needing to call back into the main r thread. this offers the best performance. if the value is \code{TRUE}, then instead of sending a 404 response, httpserver will call the application's \code{call} function, and allow it to handle the request.}

\item{html_charset}{when html files are served, the value that will be provided for \code{charset} in the content-type header. for example, with the default value, \code{"utf-8"}, the header is \code{content-type: text/html; charset=utf-8}. if \code{""} is used, then no \code{charset} will be added in the content-type header.}

\item{headers}{additional headers and values that will be included in the response.}

\item{validation}{an optional validation pattern. presently, the only type of validation supported is an exact string match of a header. for example, if \code{validation} is \code{'"abc" = "xyz"'}, then http requests must have a header named \code{abc} (case-insensitive) with the value \code{xyz} (case-sensitive). if a request does not have a matching header, than httpserver will give a 403 forbidden response. if the \code{character(0)} (the default), then no validation check will be performed.}

\item{exclude}{should this path be excluded from static serving? (this is only to be used internally, for \code{\link[=exclude_static_path]{exclude_static_path()}}.)}

\item{x}{a \code{static_path_options} object.}

\item{...}{further arguments passed to or from other methods (currently unused).}
}

\value{
an object of class \code{static_path_options}, containing the normalized options used when serving static paths.
}

\description{
create options for static paths
}

