% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stop_all_servers}
\alias{stop_all_servers}
\title{stop all servers}
\usage{
stop_all_servers()
}

\value{
no return value, called for its side effect of stopping all running servers.
}

\description{
this will stop all applications which were created by \code{\link[=start_server]{start_server()}} or \code{\link[=start_pipe_server]{start_pipe_server()}}.
}

\seealso{
\code{\link[=stop_server]{stop_server()}} to stop a specific server.
}

