% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stop_server}
\alias{stop_server}
\title{stop a server}
\usage{
stop_server(server)
}

\arguments{
\item{server}{a server object that was previously returned from \code{\link[=start_server]{start_server()}} or \code{\link[=start_pipe_server]{start_pipe_server()}}.}
}

\value{
no return value, called for its side effect of stopping the server.
}

\description{
given a server object that was returned from a previous invocation of \code{\link[=start_server]{start_server()}} or \code{\link[=start_pipe_server]{start_pipe_server()}}, this closes all open connections for that server and unbinds the port.
}

\seealso{
\code{\link[=stop_all_servers]{stop_all_servers()}} to stop all servers.
}

