% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stopDaemonizedServer}
\alias{stopDaemonizedServer}
\title{Stop a running daemonized server in Unix environments (deprecated)}
\arguments{
\item{server}{A server object that was previously returned from
  [startServer()] or [startPipeServer()].}
}

\description{
This function will be removed in a future release of httpuv. Instead, use
[stopServer()].
}

