% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{WebServer}
\alias{WebServer}
\title{WebServer class}
\description{
This class represents a web server running one application. Multiple servers
can be running at the same time.
}

\seealso{
[Server()] and [PipeServer()].
}

\keyword{internal}
