% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{as.staticPath}
\alias{as.staticPath}
\alias{as.staticPath.staticPath}
\alias{as.staticPath.character}
\alias{as.staticPath.default}
\title{Convert an object to a staticPath object}
\usage{
as.staticPath(path)
as.staticPath.staticPath(path)
as.staticPath.character(path)
as.staticPath.default(path)
}

\arguments{
\item{path}{The object to convert.}
}

\description{
This function is used to convert an object to a `staticPath` object.
It is primarily used internally, but can also be used by users to ensure
that an object is a `staticPath` object.
}

\keyword{internal}
