% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{listServers}
\alias{listServers}
\title{List all running httpserver servers}
\usage{
listServers()
}

\value{
A list of currently running server objects. The list is empty when no servers are running.
}

\description{
This returns a list of all running httpserver server applications.
}

