% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in utils.R
\name{logLevel}
\alias{logLevel}
\title{Get and set logging level}
\usage{
logLevel(level = NULL)
}

\arguments{
\item{level}{The logging level. Must be one of `NULL`, `"OFF"`,
  `"ERROR"`, `"WARN"`, `"INFO"`, or `"DEBUG"`. If
  `NULL` (the default), then this function simply returns the current
  logging level.}
}

\value{
If `level=NULL`, then this returns the current logging level. If
  `level` is any other value, then this returns the previous logging
  level, from before it is set to the new value.
}

\description{
The logging level for httpserver can be set to report differing levels of
information. Possible logging levels (from least to most information
reported) are: `"OFF"`, `"ERROR"`, `"WARN"`, `"INFO"`, or
`"DEBUG"`. The default level is `ERROR`.
}

\keyword{internal}
