% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in staticServer.R
\name{runStaticServer}
\alias{runStaticServer}
\title{Serve a directory}
\usage{
runStaticServer(
  dir = getwd(),
  host = "127.0.0.1",
  port = NULL,
  ...,
  background = FALSE,
  browse = interactive()
)
}

\arguments{
\item{dir}{The directory to serve. Defaults to the current working directory.}

\item{background}{Whether to run the server in the background. By default,
  the server runs in the foreground and blocks the R console. You can stop
  the server by interrupting it with `Ctrl + C`.

  When `background = TRUE`, the server will run in the background and will
  process requests when the R console is idle. To stop a background server,
  call [stopAllServers()] or call [stopServer()] on the server object
  returned (invisibly) by this function.}

\item{browse}{Whether to automatically open the served directory in a web
  browser. Defaults to `TRUE` when running interactively.}

\item{host}{A string that is a valid IPv4 address that is owned by this
  server, or `"0.0.0.0"` to listen on all IP addresses.}

\item{port}{A number or integer that indicates the server port that should be
  listened on. Note that on most Unix-like systems including Linux and macOS,
  port numbers smaller than 1024 require root privileges.}
}

\value{
Starts a server on the specified host and port. By default the
  server runs in the foreground and is accessible at `http://127.0.0.1:7446`.
  When `background = TRUE`, the `server` object is returned invisibly.
}

\description{
`runStaticServer()` provides a convenient interface to start a server to host
a single static directory, either in the foreground or the background.
}

\seealso{
[runServer()] provides a similar interface for running a dynamic
  app server. Both `runStaticServer()` and [runServer()] are built on top of
  [startServer()], [service()] and [stopServer()]. Learn more about httpuv
  servers in [startServer()].
}

