% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in static_paths.R
\name{staticPath}
\alias{staticPath}
\alias{excludeStaticPath}
\alias{print.staticPath}
\alias{format.staticPath}
\title{Create a staticPath object}
\usage{
staticPath(
  path,
  indexhtml = NULL,
  fallthrough = NULL,
  html_charset = NULL,
  headers = NULL,
  validation = NULL
)
excludeStaticPath()
print.staticPath(x, ...)
format.staticPath(x, ...)
}

\arguments{
\item{path}{The local path.}

\item{indexhtml}{If an index.html file is present, should it be served up
  when the client requests the static path or any subdirectory?}

\item{fallthrough}{With the default value, `FALSE`, if a request is made
  for a file that doesn't exist, then httpserver will immediately send a 404
  response from the background I/O thread, without needing to call back into
  the main R thread. This offers the best performance. If the value is
  `TRUE`, then instead of sending a 404 response, httpserver will call the
  application's `call` function, and allow it to handle the request.}

\item{html_charset}{When HTML files are served, the value that will be
  provided for `charset` in the Content-Type header. For example, with
  the default value, `"utf-8"`, the header is `Content-Type:
  text/html; charset=utf-8`. If `""` is used, then no `charset`
  will be added in the Content-Type header.}

\item{headers}{Additional headers and values that will be included in the
  response.}

\item{validation}{An optional validation pattern. Presently, the only type of
  validation supported is an exact string match of a header. For example, if
  `validation` is `'"abc" = "xyz"'`, then HTTP requests must have a
  header named `abc` (case-insensitive) with the value `xyz`
  (case-sensitive). If a request does not have a matching header, than
  httpserver will give a 403 Forbidden response. If the `character(0)` (the
  default), then no validation check will be performed.}
}

\description{
The `staticPath` function creates a `staticPath` object. Note that
if any of the arguments (other than `path`) are `NULL`, then that
means that for this particular static path, it should inherit the behavior
from the staticPathOptions set for the application as a whole.
}

\details{
The `excludeStaticPath` function tells the application to ignore a
particular path for static serving. This is useful when you want to include a
path for static serving (like `"/"`) but then exclude a subdirectory of
it (like `"/dynamic"`) so that the subdirectory will always be passed to
the R code for handling requests. `excludeStaticPath` can be used not
only for directories; it can also exclude specific files.
}

\seealso{
[staticPathOptions()].
}

