% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stopAllServers}
\alias{stopAllServers}
\title{Stop all servers}
\usage{
stopAllServers()
}

\value{
No return value, called for its side effect of stopping all running servers.
}

\description{
This will stop all applications which were created by \code{\link[=startServer]{startServer()}} or \code{\link[=startPipeServer]{startPipeServer()}}.
}

\seealso{
\code{\link[=stopServer]{stopServer()}} to stop a specific server.
}

