% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stopAllServers}
\alias{stopAllServers}
\title{Stop all servers}
\usage{
stopAllServers()
}

\description{
This will stop all applications which were created by
[startServer()] or [startPipeServer()].
}

\seealso{
[stopServer()] to stop a specific server.
}

