% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in server.R
\name{stopServer}
\alias{stopServer}
\title{Stop a server}
\usage{
stopServer(server)
}

\arguments{
\item{server}{A server object that was previously returned from
  [startServer()] or [startPipeServer()].}
}

\description{
Given a server object that was returned from a previous invocation of
[startServer()] or [startPipeServer()], this closes all
open connections for that server and unbinds the port.
}

\seealso{
[stopAllServers()] to stop all servers.
}

