% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in socket.R
\name{web_socket}
\alias{web_socket}
\title{web_socket class}
\usage{
web_socket(handle, req)
}

\arguments{
\item{handle}{an c++ web_socket handle.}

\item{req}{the rook request environment that opened the connection.}
}

\value{
an environment of class \code{web_socket} containing the connection handle, request information, callback registries, and methods for sending messages or closing the connection.
}

\description{
a \code{web_socket} object represents a single web_socket connection. the object can be used to send messages and close the connection, and to receive notifications when messages are received or the connection is closed.
}

\details{
This server-side class is different from the client-side connection provided by the package named \code{websocket}.

web_socket objects should never be created directly. they are obtained by passing an \code{on_wsopen} function to \code{\link[=start_server]{start_server()}}.
}

\examples{
# a web_socket echo server that listens on port 8080
s <- start_server(
  "0.0.0.0", 8080,
  list(
    on_headers = function(req) {
      # print connection headers
      cat(capture.output(str(as.list(req))), sep = "\n")
    },
    on_wsopen = function(ws) {
      cat("connection opened.\n")

      ws$on_message(function(binary, message) {
        cat("server received message:", message, "\n")
        ws$send(message)
      })
      ws$on_close(function() {
        cat("connection closed.\n")
      })
    }
  )
)

s$stop()
}

