% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in socket.R
\name{WebSocket}
\alias{WebSocket}
\title{WebSocket class}
\usage{
WebSocket(handle, req)
}

\arguments{
\item{handle}{An C++ WebSocket handle.}

\item{req}{The Rook request environment that opened the connection.}
}

\value{
An environment of class \code{WebSocket} containing the connection handle, request information, callback registries, and methods for sending messages or closing the connection.
}

\description{
A \code{WebSocket} object represents a single WebSocket connection. The object can be used to send messages and close the connection, and to receive notifications when messages are received or the connection is closed.
}

\details{
Note that this WebSocket class is different from the one provided by the package named websocket. This class is meant to be used on the server side, whereas the one in the websocket package is to be used as a client. The WebSocket class in httpserver has an older API than the one in the websocket package.

WebSocket objects should never be created directly. They are obtained by passing an \code{onWSOpen} function to \code{\link[=startServer]{startServer()}}.
}

\examples{
# A WebSocket echo server that listens on port 8080
s <- startServer(
  "0.0.0.0", 8080,
  list(
    onHeaders = function(req) {
      # Print connection headers
      cat(capture.output(str(as.list(req))), sep = "\n")
    },
    onWSOpen = function(ws) {
      cat("Connection opened.\n")

      ws$onMessage(function(binary, message) {
        cat("Server received message:", message, "\n")
        ws$send(message)
      })
      ws$onClose(function() {
        cat("Connection closed.\n")
      })
    }
  )
)

s$stop()
}

